create a memoized profile items
